+++ config/sanctum.php
@@ -22,7 +22,7 @@ return [
 
     'stateful' => explode(',', env('SANCTUM_STATEFUL_DOMAINS', sprintf(
         '%s%s',
-        'localhost,localhost:3000,127.0.0.1,127.0.0.1:8000,::1',
+        'localhost,localhost:3000,localhost:5173,127.0.0.1,127.0.0.1:8000,127.0.0.1:5173,::1',
         Sanctum::currentApplicationUrlWithPort(),
         // Sanctum::currentRequestHost(),
     ))),
@@ -52,7 +52,9 @@ return [
     |
     */
 
-    'expiration' => null,
+    'expiration' => (($days = config('api.token_expiration_days')) > 0)
+        ? $days * 24 * 60
+        : null,
 
     /*
     |--------------------------------------------------------------------------